# IndianMasterApp-Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta:
        description: Pagination metadata or other metadata
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@indianmaster.com
    name: API Support
  description: Backend APIs for Indian Master Application - A comprehensive platform
    for job seekers and hirers in India
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Indian Master API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of users filtered by role
      parameters:
      - description: User role to filter by
        in: query
        name: role
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g., created_at, full_name)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully with pagination metadata
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch users
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Users by Role
      tags:
      - User
  /users/active:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of all active users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Sort field (e.g., created_at, full_name)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Active users retrieved successfully with pagination metadata
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to fetch active users
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Active Users
      tags:
      - User
  /worker/{worker_id}/recommended-jobs:
    get:
      consumes:
      - application/json
      description: Retrieve job recommendations for a worker based on matching criteria.
        Results are cached in Redis for 5 minutes.
      parameters:
      - description: Worker ID
        in: path
        name: worker_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, job_role)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job recommendations retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to get recommended jobs
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Job Recommendations (Cached)
      tags:
      - Worker
...
  /workers/{worker_id}/recommended-jobs:
    get:
      consumes:
      - application/json
      description: Retrieve job recommendations for a worker based on matching criteria
      parameters:
      - description: Worker ID
        in: path
        name: worker_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: match_score
        description: Sort field (e.g., match_score, job_role)
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order (asc or desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job recommendations retrieved successfully
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Failed to get recommended jobs
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Get Job Recommendations
      tags:
      - Worker
...
schemes:
- http
- https
swagger: "2.0"